errorsReducer now handles listener errors. pathWithMeta moved to utils.
