+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "redux-firestore",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "Redux bindings for Firestore.",
   "main": "lib/index.js",
   "module": "es/index.js",
